+++ src/retest/__init__.py
@@ -1,8 +1,12 @@
 __version__ = "0.1.0"
 
-from ._test import test, run_tests
+from ._decorator import test
+from ._runner import run_tests
+from ._types import TestFunc, TestRunner
 
 __all__ = [
     "test",
     "run_tests",
+    "TestFunc",
+    "TestRunner",
 ]
