+++ src/retest/_types.py
@@ -1,0 +1,4 @@
+from typing import Callable
+
+type TestFunc = Callable[[], None]
+type TestRunner = Callable[[], bool]
